modules/persons_service/requirements.txt
GeoAlchemy2==0.8.4
greenlet==2.0.2
grpcio==1.54.2
protobuf==4.23.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Shapely==1.7.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
